Fix building on Windows

+++ Sources/Gtk/CssProvider.swift
@@ -9,7 +9,7 @@ public class CssProvider {
 
     public func loadFromData(_ data: String) {
         data.withCString { cString in
-            gtk_css_provider_load_from_data(pointer, cString, data.count)
+            gtk_css_provider_load_from_data(pointer, cString, gssize(data.count))
         }
     }
 }
